refactor matching
