+++ internal/matcher/internal/body/range.go
@@ -3,7 +3,7 @@ package body
 import (
 	"fmt"
 
-	"github.com/g8rswimmer/http-loki/internal/mock/internal/matcher/internal/validate"
+	"github.com/g8rswimmer/http-loki/internal/matcher/internal/validate"
 	"github.com/g8rswimmer/http-loki/internal/model"
 	"github.com/tidwall/gjson"
 )
