+++ internal/matcher/internal/body/regex.go
@@ -1,7 +1,7 @@
 package body
 
 import (
-	"github.com/g8rswimmer/http-loki/internal/mock/internal/matcher/internal/validate"
+	"github.com/g8rswimmer/http-loki/internal/matcher/internal/validate"
 	"github.com/g8rswimmer/http-loki/internal/model"
 	"github.com/tidwall/gjson"
 )
