+++ internal/matcher/internal/query/regex_test.go
